requirements.txt
alembic==1.0.8
argh==0.26.2
Click==7.0
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.8
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
pathtools==0.1.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
PyYAML==5.1
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.1
watchdog==0.9.0
Werkzeug==0.15.1
